Restore assemble_matrix.cpp from jorgensd/main
Motivation
==========
Jorgen pointed out that using `conj` only for complex builds can be
achieved using `if constexpr`. This eliminates the need for separate
real-only and complex-only implementations of `modify_mpc_cell_rows`
and `modify_mpc_cell_masters`. The required functionality can be
easily implemented by adding `if constexpr` in to the previous
implementation of `modify_mpc_cell`; so there is no need for the
methods `modify_mpc_cell_rows` and `modify_mpc_cell_masters`.

Changes
=======
This commit restores `assemble_matrix.cpp` to the state that existed
on jorgensd/dolfinx_mpc/main before development was started on the
hermitian transpose branch.